requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==45.0.6
cssselect2==0.8.0
curl_cffi==0.13.0
Django==5.2.5
django-cloudinary-storage==0.3.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
fpdf==1.7.2
frozendict==2.4.6
geographiclib==2.1
geopy==2.4.1
gevent==25.9.1
geventhttpclient==2.3.7
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mercadopago==2.3.0
msal==1.33.0
msgpack==1.1.2
msgspec==0.19.0
multitasking==0.0.12
mysql-connector==2.2.9
mysql-connector-python==9.4.0
mysqlclient==2.2.7
ntplib==0.4.0
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
panda==0.3.1
pandas==2.3.1
paypalrestsdk==1.13.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.3.1
pyphen==0.17.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2024.1
pywin32==311
pyzmq==27.1.0
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
transbank-sdk==6.1.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.3
wsproto==1.3.2
yfinance==0.2.65
zope.event==6.1
zope.interface==8.1.1
zopfli==0.4.0
